Update frontend migration logic
- Add version to all queries
- Move migration to own file with tests
