+++ src/utils/version.ts
@@ -1,4 +1,7 @@
+import pluginPackage from '../../package.json';
 import { isNumber } from 'lodash';
+
+export const pluginVersion = pluginPackage.version;
 
 const versionPattern = /^(\d+)(?:\.(\d+))?(?:\.(\d+))?(?:-([0-9A-Za-z\.]+))?/;
